complete fsengine crud

+++ configs/configs.go
@@ -3,19 +3,19 @@ package configs
 import "sync"
 
 var configs = struct {
-	MaxSizeVirtualFile uint32
+	MaxSizeVirtualFile uint64
 	mutex sync.Mutex
 }{
 	MaxSizeVirtualFile: 10*1024*1024*1024,  // 10 GB
 }
 
-func SetConfigs(maxSizeVirtualFile uint32) {
+func SetConfigs(maxSizeVirtualFile uint64) {
 	configs.mutex.Lock()
 	defer configs.mutex.Unlock()
 	configs.MaxSizeVirtualFile = maxSizeVirtualFile
 }
 
-func GetMaxSizeVirtualFile() uint32 {
+func GetMaxSizeVirtualFile() uint64 {
 	configs.mutex.Lock()
 	defer configs.mutex.Unlock()
 	return configs.MaxSizeVirtualFile
